config macro
Summary:
this adds a config macro for parsing the env var into a struct.

there are 3 attributes
#[env(ENVVAR1,ENVVAR2,...)]
this will determine what enn var are used to set the field. If more than one env var is non-empty,
the first env var in the list is used.
#[default("/etc/logdna/config.yaml")]
this sets a default value if all the env vars in the list are empty. Fields without a default must
be Option<T>
#[example("/etc/logdna/config.yaml")]
this is used for generating unit test data but it will aslo be used to generate docs in the future.

Test Plan: It complies and i have verfied the geenerated code for correctness

Reviewers: #rust, dchai

Reviewed By: #rust, dchai

Differential Revision: https://logdna.phacility.com/D1247