middleware draft
Summary:
Middlewares are designed to mutate lines produced by the tailer.
The trait contains an init() which is called once when the executor is started.
It also contains a fn process(&self, line: LineBuilder) -> Result<LineBuilder, LineBuilder>.
This takes in a line and returns Ok if the line should be moved to the next step and Err if it should be skipped.
Although there are no plans to skip lines I see it as good future proofing.

Test Plan: N/a

Reviewers: #rust, dchai

Reviewed By: #rust, dchai

Differential Revision: https://logdna.phacility.com/D1324